Fix Navigate Plane pinch zoom, add pinch rotate, and move cube
The Navigate Plane scene is complete. On start, a board made of a
16-tile perimeter is placed on the closest detected AR plane. On
touch and hold at a board tile, a cube is placed on the tile
center. Dragging the cube or touching and holding on another tile
will move the cube around the tiles of the perimeter. The board
can be pulled towards the camera or pushed further away with
pinch zoom. Pinch rotate will rotate the board around its center.

Changes:
 * Implemented "debouncing" control for distinguishing different
   haptics.
 * Zoom is done by +/- the
   m_SessionOrigin.transform.localPosition
 * The Plane is instantiated on a detected AR Plane.
 * Plane rotation is done by +/- the
   plane.transform.rotation.eulerAngles.y
 * The Cube is instantiated and moved on ray traced objects in
   the scene, i.e. the board tiles.