Add Hand Gestures scene and Charades plugin
Hand Gestures scene depends on Charades plugin, which has been added in
the project. The plugin itself relies on Charades framework, which
depends on MediaPipe. The framework is not added to the repository. It
is build in the MediaPipe workspace and copied to the plugin directory.
The MediaPipe workspace is a git submodule of the MediaPipe SDK, which
is added as a submodule to this repository. The Makefile has been
updated to build the Charades framework, install it in the Unity plugin
directory and then build the Unity iOS project.

Summary of changes:
 * Added MediaPipe SDK submodule, which includes the MediaPipe
   workspace.
 * Added Charades plugin responsible for hand tracking.
 * Updated the Makefile to build and install the Charades framework as
   a dependency of the Unity project.